% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afm_pisa_heatmap.R
\name{afm.pisa.heatmap}
\alias{afm.pisa.heatmap}
\title{Heatmap for PAE, deltaG, interfaceArea and surfaceArea for predicted AFM models}
\usage{
afm.pisa.heatmap(import_afm, import_pisa)
}
\arguments{
\item{import_pisa}{}
}
\value{
ggplot geom_tile
}
\description{
Heatmap for PAE, deltaG, interfaceArea and surfaceArea for predicted AFM models
}
