% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_afm.R
\name{import.afm}
\alias{import.afm}
\title{Import PAE, pLLDT and PTM from AFM predictions .json files}
\usage{
import.afm(dir = "data/")
}
\arguments{
\item{dir:}{specify directory where '*scores.json' files are stored}
}
\value{
list containing elements
}
\description{
Import PAE, pLLDT and PTM from AFM predictions .json files
}
\examples{
afm_json <- import.afm()
}
