% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pisa.R
\name{import.pisa}
\alias{import.pisa}
\title{Import interface summary from PISA .xml files}
\usage{
import.pisa(dir = "data")
}
\arguments{
\item{dir:}{directory where .xml files are saved}
}
\value{
data.frame
}
\description{
Import interface summary from PISA .xml files
}
\examples{
import_pisa <- import.pisa(dir="data")
}
