% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pae_boxplot.R
\name{pae.boxplot}
\alias{pae.boxplot}
\title{Boxplot to visualize PAE for all models of a AFM predicted complex}
\usage{
pae.boxplot(pae_interface)
}
\arguments{
\item{pae_interface}{}
}
\value{
boxplot
}
\description{
Boxplot to visualize PAE for all models of a AFM predicted complex
}
