% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pae_heatmap.R
\name{pae.heatmap}
\alias{pae.heatmap}
\title{Plot kmeans interface clusters}
\usage{
pae.heatmap(
  import_afm,
  afm_complex = "NSP14-NSP10",
  afm_rank = 1,
  interface_cluster = FALSE,
  plldt = 50
)
}
\arguments{
\item{import_afm:}{list in the format from import.afm()}

\item{afm_complex:}{AFM predicted complex in the format "NSP10-NSP16"}

\item{afm_rank:}{which afm_rank from AFM predictions to plot}

\item{interface_cluster:}{cluster PAE values}

\item{plldt:}{pLLDT cutoff}
}
\value{
ComplexHeatmap
}
\description{
Plot kmeans interface clusters
}
