% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pae_interface.R
\name{pae.interface}
\alias{pae.interface}
\title{Perform kmeans clustering to identify amino acid cluster with the lowest PAE}
\usage{
pae.interface(import_afm, plldt = 50)
}
\arguments{
\item{afm:}{list; result from import.afm() function}

\item{plldt:}{pLLDT cutoff}
}
\value{
data frame
}
\description{
Perform kmeans clustering to identify amino acid cluster with the lowest PAE
}
\examples{
import_afm <- import.afm()
pae_interface <- pae.interface()
}
