% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa_boxplot.R
\name{pisa.boxplot}
\alias{pisa.boxplot}
\title{Boxplot of PISA results}
\usage{
pisa.boxplot(import_pisa)
}
\arguments{
\item{import_pisa}{}
}
\value{
ggplot2 boxplot
}
\description{
Boxplot of PISA results
}
