% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plldt_lineplot.R
\name{plldt.lineplot}
\alias{plldt.lineplot}
\title{Plot pLLDT as lineplot}
\usage{
plldt.lineplot(import_afm, afm_complex = "NSP10-NSP16", afm_rank = "all")
}
\arguments{
\item{import_afm:}{list in the format from import.afm()}

\item{afm_complex:}{AFM predicted complex in the format "NSP10-NSP16"}

\item{afm_rank:}{"all" or integer: which afm_rank from AFM predictions to plot}
}
\value{
ggplot
}
\description{
Plot pLLDT as lineplot
}
